migrate

+++ src/main/java/com/example/accountdemo/api/dto/AccountBalanceItem.java
@@ -1,5 +1,7 @@
 package com.example.accountdemo.api.dto;
 
+import java.math.BigDecimal;
+
 /** Một dòng số dư trong response xem ví. */
-public record AccountBalanceItem(String currency, long available, long locked) {
+public record AccountBalanceItem(String currency, BigDecimal available, BigDecimal locked) {
 }
