+++ src/main/java/com/example/accountdemo/api/dto/AmountRequest.java
@@ -1,5 +1,7 @@
 package com.example.accountdemo.api.dto;
 
+import java.math.BigDecimal;
+
 /** Body nạp / rút tiền. */
-public record AmountRequest(long amount, String currency) {
+public record AmountRequest(BigDecimal amount, String currency) {
 }
