+++ src/main/java/com/example/accountdemo/api/dto/HoldingRequestItem.java
@@ -1,5 +1,7 @@
 package com.example.accountdemo.api.dto;
 
+import java.math.BigDecimal;
+
 /** Một dòng số dư trong body tạo account (locked = 0 ở Application). */
-public record HoldingRequestItem(String currency, long available) {
+public record HoldingRequestItem(String currency, BigDecimal available) {
 }
