+++ src/main/java/com/example/accountdemo/api/dto/PlaceOrderRequest.java
@@ -1,9 +1,11 @@
 package com.example.accountdemo.api.dto;
 
+import java.math.BigDecimal;
+
 /**
  * Body đặt lệnh: side, type, cặp, quantity, price (null nếu MARKET).
  *
- * <p><b>Vì sao cần class này:</b> nhận input HTTP dạng string/long trước khi map sang value object domain.
+ * <p><b>Vì sao cần class này:</b> nhận input HTTP dạng string/number trước khi map sang value object domain.
  */
 public record PlaceOrderRequest(
         String accountId,
@@ -11,7 +13,7 @@ public record PlaceOrderRequest(
         String orderType,
         String baseCurrency,
         String quoteCurrency,
-        long quantity,
-        Long price
+        BigDecimal quantity,
+        BigDecimal price
 ) {
 }
